// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-emergency-contact-details-model.ts
export class EmployeeEmergencyContactDetails {
    employeeEmergencyContactId: string;
    employeeId: string;
    relationshipId: string;
    surName: string;
    firstName: string;
    lastName: string;
    otherRelation: string;
    homeTelephone: string;
    mobileNo: Date;
    expiryDate: Date;
    workTelephone: string;
    isEmergencyContact: boolean;
    addressStreetOne: string;
    addressStreetTwo:string;
    isDependentContact:boolean;
    zipOrPostalCode:string;
    countryId: string;
    isArchived:boolean;
    id:string;
    timeStamp:any;
    emergencyContactId:string;
    totalCount: number;
}